backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-colorfield==0.11.0
djoser==2.2.2
drf-extra-fields==3.7.0
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
PyYAML==6.0
